IVK-403 added list of offending accessions to "internal error"
When there are duplicate accessions in the VMR, the merge code betwee the blast results tsv and the VMR metadata creates extra rows, and triggers an internal error.

That error now includes the row counts, and the actual offending accessions, to help in debugging.

It remains an error, though it would likely be better for the user to process the data and issue a warning, but we have no channel for warnings, so would never know about it. We need to add a duplicate check to the makedatabase step to prevent this.